[chore] #41 충돌 해결 후 TemplateQueryService 구조 통합
